Changed lat and long types to double

+++ src/WebApplication3/Models/PhoneSpotting.cs
@@ -13,8 +13,8 @@ namespace WebApplication3.Models
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int PhoneSpottingId { get; set; }
         public DateTime SpotTime { get; set; }
-        public int Longitude { get; set; }
-        public int Latitude { get; set; }
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
         public string PhoneManufacturer { get; set; }
         public string PhoneModel { get; set; }
         public string Notes { get; set; }
